layer_norm_fix default axes is 0 clean unused headers
